refactor(archive): move artifact taxonomy out of lib
