+++ devtools/build_topology_projection.py
@@ -89,8 +89,8 @@ LIB_PREFIX_TO_SUBPACKAGE = {
     "viewports": "lib/viewport/",
     "raw_payload_": "lib/raw_payload/",
     "raw_payload": "lib/raw_payload/",
-    "artifact_taxonomy_": "lib/artifact_taxonomy/",
-    "artifact_taxonomy": "lib/artifact_taxonomy/",
+    "artifact_taxonomy_": "archive/artifact_taxonomy/",
+    "artifact_taxonomy": "archive/artifact_taxonomy/",
     "action_event_": "archive/action_event/",
     "action_events": "archive/action_event/",
     "message_": "lib/message/",
@@ -206,7 +206,7 @@ TARGET_TO_OWNER = [
     ("polylogue/lib/session/", "lib-session"),
     ("polylogue/lib/viewport/", "lib-viewport"),
     ("polylogue/lib/raw_payload/", "lib-raw-payload"),
-    ("polylogue/lib/artifact_taxonomy/", "lib-artifact-taxonomy"),
+    ("polylogue/archive/artifact_taxonomy/", "archive-artifact-taxonomy"),
     ("polylogue/archive/action_event/", "archive-action-event"),
     ("polylogue/lib/message/", "lib-message"),
     ("polylogue/lib/conversation/", "lib-conversation"),
