+++ docs/plans/topology-target.yaml
@@ -71,6 +71,22 @@ files:
     loc: 121
     target: polylogue/archive/action_event/parsing.py
     owner: stable
+  - path: polylogue/archive/artifact_taxonomy/__init__.py
+    loc: 18
+    target: polylogue/archive/artifact_taxonomy/__init__.py
+    owner: stable
+  - path: polylogue/archive/artifact_taxonomy/models.py
+    loc: 35
+    target: polylogue/archive/artifact_taxonomy/models.py
+    owner: stable
+  - path: polylogue/archive/artifact_taxonomy/runtime.py
+    loc: 187
+    target: polylogue/archive/artifact_taxonomy/runtime.py
+    owner: stable
+  - path: polylogue/archive/artifact_taxonomy/support.py
+    loc: 120
+    target: polylogue/archive/artifact_taxonomy/support.py
+    owner: stable
   - path: polylogue/archive/attachment/__init__.py
     loc: 0
     target: polylogue/archive/attachment/__init__.py
@@ -518,22 +534,6 @@ files:
     target: polylogue/lib/__init__.py
     owner: lib-root
     reason: lib-root primitive
-  - path: polylogue/lib/artifact_taxonomy/__init__.py
-    loc: 18
-    target: polylogue/lib/artifact_taxonomy/__init__.py
-    owner: stable
-  - path: polylogue/lib/artifact_taxonomy/models.py
-    loc: 35
-    target: polylogue/lib/artifact_taxonomy/models.py
-    owner: stable
-  - path: polylogue/lib/artifact_taxonomy/runtime.py
-    loc: 187
-    target: polylogue/lib/artifact_taxonomy/runtime.py
-    owner: stable
-  - path: polylogue/lib/artifact_taxonomy/support.py
-    loc: 120
-    target: polylogue/lib/artifact_taxonomy/support.py
-    owner: stable
   - path: polylogue/lib/conversation/__init__.py
     loc: 0
     target: polylogue/lib/conversation/__init__.py
