+++ polylogue/pipeline/services/ingest_worker.py
@@ -18,7 +18,7 @@ from typing import TYPE_CHECKING, Literal
 
 from typing_extensions import TypedDict
 
-from polylogue.lib.artifact_taxonomy import ArtifactClassification, classify_artifact
+from polylogue.archive.artifact_taxonomy import ArtifactClassification, classify_artifact
 from polylogue.lib.conversation.branch_type import BranchType
 from polylogue.lib.json import dumps as json_dumps
 from polylogue.lib.raw_payload.decode import RawPayloadEnvelope
