+++ polylogue/schemas/observation_runtime.py
@@ -7,7 +7,7 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import TypeAlias
 
-from polylogue.lib.artifact_taxonomy import classify_artifact
+from polylogue.archive.artifact_taxonomy import classify_artifact
 from polylogue.lib.json import JSONDocument, JSONValue, is_json_value, json_document
 from polylogue.lib.raw_payload import extract_payload_samples, record_bucket_key
 from polylogue.schemas.observation_identity import derive_bundle_scope, schema_cluster_id
