+++ polylogue/sources/decoder_zip.py
@@ -6,7 +6,7 @@ import zipfile
 from collections.abc import Iterable
 from pathlib import Path
 
-from polylogue.lib.artifact_taxonomy import classify_artifact_path
+from polylogue.archive.artifact_taxonomy import classify_artifact_path
 from polylogue.logging import get_logger
 from polylogue.storage.blob_store import get_blob_store
 from polylogue.storage.cursor_state import CursorStatePayload
